requirements.txt
discord.py>=2.3.0
alpaca-py>=0.18.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytz>=2024.1
aiohttp>=3.8.0
discord.py
Flask
requests
yfinance
pandas
numpy
